refactor: move utility components to utils folder
