+++ src/components/Item/tests/Item.test.tsx
@@ -1,6 +1,6 @@
 import '@testing-library/jest-dom';
 import { fireEvent, render, screen } from '@testing-library/react';
-import { server } from '../../MockService/Server';
+import { server } from '../../../utils/MockService/Server';
 import { App } from '../../../App';
 import { queryItem } from '../../API/API';
 import { IBeerDetails, SEARCH_DEFAULT } from '../../../shared/data/data';
